database/requirements.txt
faker
numpy
pyodbc
mysql-connector-python
